==> 2.ETLcredits.ipynb <==
EXTRACCIÓN, TRANSFORMACIÓN, CARGA DE DATOS (E.T.L)


María Agostina Balverde Moll - DATA FT 11

#Importar librerías necesarias

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import csv
import json
import ast

#Importo el csv 'credits'

data_credits = pd.read_csv('Datasets/credits.csv')

data_credits['cast'] = data_credits['cast'].apply(ast.literal_eval)

#El campo 'cast' anidado, lo desanido y coloco en nuevas columnas los datos que contiene

data_credits['CastId'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['cast_id']) for d in x]))

data_credits['Character'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['character']) for d in x]))

data_credits['CreditId'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['credit_id']) for d in x]))

data_credits['Gender'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['gender']) for d in x]))

data_credits['Id'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['id']) for d in x]))

data_credits['Name'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['name']) for d in x]))

data_credits['Order'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['order']) for d in x]))

data_credits['ProfilePath'] = data_credits['cast'].apply(lambda x: ', '.join([str(d['profile_path']) for d in x]))

#Elimino la columna 'cast', 'CastId', 'CreditId', 'Gender', 'Id', 'Order', 'ProfilePath'

data_credits = data_credits.drop('cast', axis=1)
data_credits = data_credits.drop('CastId', axis=1)
data_credits = data_credits.drop('CreditId', axis=1)
data_credits = data_credits.drop('Gender', axis=1)
data_credits = data_credits.drop('Id', axis=1)
data_credits = data_credits.drop('Order', axis=1)
data_credits = data_credits.drop('ProfilePath', axis=1)

#Cargo los datos en el csv 'credits'

data_credits.to_csv('Datasets/credits.csv', index=False)

data_credits.head(5)

#Creo un nuevo dataframe para trabajar con la columna 'crew'

data_credits_1 = pd.read_csv('Datasets/credits.csv')

#Veo la primera fila del dataframe

data_credits_1 ['crew'] [0]

#Creo una función para obtener sólo el nombre de los directores de las películas de la columna 'crew'

def obtener_director (obj):
  lista_director = []
  for i in ast.literal_eval(obj):
    if i['job'] == 'Director':
      lista_director.append(i['name'])
      break
  return lista_director

#Aplico la función obtener_director

data_credits_1['crew'] = data_credits_1['crew'].apply(obtener_director)

#Cambio el nombre de la columna 'crew' por 'Director'

#Creo un diccionario con los nuevos nombres de las columnas
nuevos_nombres = {
    'crew': 'Director',
    'id': 'Id',
}

#Cambio los nombres de las columnas
data_credits_1 = data_credits_1.rename(columns=nuevos_nombres)

data_credits_1.head(5)

#Cargo los datos en el csv 'credits'

data_credits_1.to_csv('Datasets/credits.csv', index=False)
