==> 3.ETLunircsv.ipynb <==
EXTRACCIÓN, TRANSFORMACIÓN, CARGA DE DATOS (E.T.L)


María Agostina Balverde Moll - DATA FT 11

#Importo la librería pandas

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#Cargo en dos datafrarmes los csv modificados en los pasos anteriores

data_movies = pd.read_csv('datos_modificados.csv')
data_credits = pd.read_csv('Datasets/credits.csv')

#Reviso el tipo de dato de la columna 'id' de ambos dataframe

print(data_movies['id'].dtype)

print(data_credits['Id'].dtype)

#Filtro la columa 'id' para buscar los valores que no son enteros

noints = data_movies[~data_movies['id'].astype(str).str.isdigit()].reset_index(drop=True)

columna_noints = data_credits.loc[noints.index, 'Id']

print(columna_noints)

import pandas as pd

# Filtrar donde la columna 'id' no es de tipo int
noints = data_movies[~data_movies['id'].astype(str).str.isdigit()]

# Convertir los datos de la columna 'id' que no son de tipo entero a NaN
data_movies.loc[noints.index, 'id'] = pd.np.nan

# Convertir la columna 'id' a tipo entero
data_movies['id'] = data_movies['id'].astype(float).astype(pd.Int64Dtype())

# Mostrar los datos convertidos a tipo entero
print(data_movies['id'])

data_merge = pd.merge (data_movies, data_credits, how = 'outer', left_index=True, right_index=True)

data_merge.head(0)

#Elimino las columnas que no son necesarias a los fines de crear las funciones

data_merge = data_merge.drop(['status', 'CollectionId','CollectionName','CollectionPosterPath','CollectionBackdropPath','spoken_languages'], axis=1)

#Elimino id que está repetida
data_merge = data_merge.drop(['id'], axis=1)

#Acomodo los nombres de las columnas

data_merge = data_merge.rename(columns={'budget': 'Budget',
                                        'genres':'Genres',
                                        'original_language':'OriginalLanguage',
                                        'overview':'Overview',
                                        'popularity':'Popularity',
                                        'production_companies':'ProductionCompanies',
                                        'production_countries':'ProductionCountries',
                                        'release_date':'ReleaseDate',
                                        'revenue':'Revenue',
                                        'runtime':'Runtime',
                                        'tagline':'Tagline',
                                        'title':'Title',
                                        'vote_average':'VoteAverage',
                                        'vote_count':'VoteCount',
                                        'release_year':'ReleaseYear',
                                        'return':'Return',
                                        'release_month':'ReleaseMonth',
                                        'Character':'MovieCharacter',
                                        'Name':'ActorName'})

data_merge.shape

#Verifico los tipos de datos

data_merge.dtypes

#'Popularity' tiene formato de objeto. Lo paso a float

#Identifico los valores no numéricos
data_merge['Popularity'] = pd.to_numeric(data_merge['Popularity'], errors='coerce')

#Convierto la columna a tipo float
data_merge['Popularity'] = data_merge['Popularity'].astype(float)


#Paso 'ReleaseDate' a datetime

data_merge['ReleaseDate'] = pd.to_datetime(data_merge['ReleaseDate'], errors='coerce')

#Extraigo sólo la fecha (sin la hora)
data_merge['ReleaseDate'] = data_merge['ReleaseDate'].dt.date

#Verifico los tipos de datos

data_merge.dtypes

data_merge.head(1)

#Verifico si hay valores duplicados

print(f"Hay {data_merge.duplicated().sum()} valores duplicados")

#Exporto el dataframe completo a un csv nuevo

data_merge.to_csv('Movies_completo.csv', index=False)
